requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.16.5
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   starlette
    #   watchfiles
click==8.2.1
    # via uvicorn
fastapi==0.116.1
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via anyio
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via mako
psycopg2-binary==2.9.10
    # via -r requirements.in
pydantic[dotenv]==2.11.9
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.1
    # via uvicorn
pyyaml==6.0.2
    # via uvicorn
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   -r requirements.in
    #   alembic
starlette==0.47.3
    # via fastapi
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
uvicorn[standard]==0.35.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
